requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
motor==3.5.2
pymongo==4.8.0
huggingface_hub==0.23.0
python-dotenv==1.0.1
jinja2==3.1.3